Adding Pages to PNGs splitter
_Note: this commit assumes that we've merged #169._

It does the following things:

1. Extracts a `BaseSplitter` class
2. Creates a `PagesToTiffsSplitter`
3. Creates a `PagesToPngsSplitter`
4. Renames the default configured splitter class to preserve past behavior.

This will allow us to then configure which splitter we use:

1. Favor TIFFs when you want deeper zooms
2. Favor PNGs when you want to minimize storage space

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/293
- https://github.com/scientist-softserv/iiif_print/pull/169
