+++ spec/iiif_print/split_pdfs/pages_to_tiffs_splitter_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require 'misc_shared'
 
-RSpec.describe IiifPrint::SplitPdfs::PagesIntoImagesService do
+RSpec.describe IiifPrint::SplitPdfs::PagesToTiffsSplitter do
   # TODO: add specs
 end
