+++ spec/iiif_print_spec.rb
@@ -35,7 +35,7 @@ RSpec.describe IiifPrint do
       end
 
       it "has a #pdf_splitter_service" do
-        expect(record.iiif_print_config.pdf_splitter_service).to be(IiifPrint::SplitPdfs::PagesIntoImagesService)
+        expect(record.iiif_print_config.pdf_splitter_service).to be(IiifPrint::SplitPdfs::PagesToTiffsService)
       end
 
       it "has #derivative_service_plugins" do
